Aggiunta una colonna alla tabella animals. Popolata con image_ url. Aggiunto il file.csv i link attivi con le immagini. Messe nelle card le img.

+++ database/migrations/2024_07_22_110541_create_animals_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('habitat', 100);
             $table->string('geographical_distribution', 100);
             $table->smallInteger('high_max');
-            $table->smallInteger('weigh_max');
+            $table->mediumInteger('weigh_max');
             $table->text('note');
             $table->timestamps();
         });
